Address PR review: unify trade phase-error key, split trade tests, document redaction contract
- M2: handleProposeTrade/handleCounterTrade now throw the canonical
  game.invalid_phase (matching other handlers) instead of game.invalid_action.
- M9: move trade schema/action tests out of the >1k-line validation.test.ts
  into tests/unit/tradeValidation.test.ts.
- M10: document the two-path trade-offer redaction contract in
  gamePersistence.ts (broadcast strips tradeOffers and relies on GameRoom +
  toClientGameState; HTTP calls toClientGameState directly).
- Add B1 regression tests for accept/reject during insider-peek and
  disruption-nullify phases, plus an invalid_phase propose-trade assertion.
